Refactoring of control-layout component by creating sub components in form_option_list.js
