+++ covid_project/frontend/src/components/page_country/country_main_page.js
@@ -1,31 +1,25 @@
-import React, { useState, useEffect } from "react";
+import React, { useState } from "react";
 import { Container } from "react-bootstrap";
 
-import { Route, useHistory } from "react-router-dom";
+import { Route } from "react-router-dom";
 
 import { ControlLayout } from "./control_layout";
 import { Countries } from "./Countries";
 
 export const MainCountry = () => {
-  let history = useHistory();
-
   const [cumulative, setCumulative] = useState(false);
   const [label, setLabel] = useState("/ Day");
 
-  const yesterday = (d) => {
+  function yesterday(d) {
     new Date(d.setDate(d.getDate() - 1));
     return d;
-  };
+  }
   const [startDate, setStartDate] = useState(new Date("2020/01/03"));
   const [endDate, setEndDate] = useState(yesterday(new Date()));
 
   const cumulativeHandleChange = (e) => {
     setCumulative(e.target.value);
     setLabel(e.target.value === "true" ? "Total" : "/ Day");
-  };
-
-  const countryHandleChange = (e) => {
-    history.push(`/country/${e.target.value}`);
   };
 
   const changeStartDate = (date) => {
@@ -40,7 +34,7 @@ export const MainCountry = () => {
     <div>
       <ControlLayout
         cumulativeHandleChange={cumulativeHandleChange}
-        countryHandleChange={countryHandleChange}
+        selectedCumulative={cumulative}
         startDate={startDate}
         endDate={endDate}
         changeStartDate={changeStartDate}
