+++ covid_project/frontend/static/frontend/main.js.LICENSE.txt
@@ -56,6 +56,8 @@
 
 /*! export Footer [provided] [no usage info] [missing usage info prevents renaming] */
 
+/*! export FormOptionList [provided] [no usage info] [missing usage info prevents renaming] */
+
 /*! export ForwardRef [provided] [no usage info] [missing usage info prevents renaming] */
 
 /*! export ForwardRef [provided] [no usage info] [provision prevents renaming (no use info)] -> ./node_modules/react-is/cjs/react-is.production.min.js .ForwardRef */
@@ -2280,6 +2282,10 @@
 
 /*!*********************************************************!*\
   !*** ./node_modules/regexp.prototype.flags/polyfill.js ***!
+  \*********************************************************/
+
+/*!*********************************************************!*\
+  !*** ./src/components/page_country/form_option_list.js ***!
   \*********************************************************/
 
 /*!**********************************************************!*\
